1) 1.1 espositore
