+++ utils/espositore_manager.py
@@ -1,4 +1,5 @@
-from PyQt5.QtWidgets import (QInputDialog, QMessageBox, QLabel, QPushButton, QHBoxLayout, QVBoxLayout, QWidget, QGridLayout)
+from PyQt5.QtWidgets import (QInputDialog, QMessageBox, QWidget, QVBoxLayout, QLabel, QPushButton,
+                             QHBoxLayout, QGridLayout)
 from PyQt5.QtCore import Qt
 
 def add_person(app):
@@ -7,6 +8,7 @@ def add_person(app):
         if ok and text:
             app.person_list.addItem(text)
             app.person_schedule[text] = {}
+            update_turni_table(app)
     except Exception as e:
         QMessageBox.critical(app, "Errore", f"Si è verificato un errore: {str(e)}")
 
@@ -25,12 +27,35 @@ def remove_person(app):
 def display_person_details(app, item):
     try:
         person_name = item.text()
+        app.current_person = person_name
+        
         if person_name in app.person_schedule:
             details = "\n".join([f"Data: {date}, Fasce Orarie: {', '.join([f['fascia_oraria'] for f in fasce])}" 
                                  for date, fasce in app.person_schedule[person_name].items()])
             app.detail_text.setText(details)
         else:
             app.detail_text.setText("Nessun dettaglio disponibile.")
+        
+        update_turni_table(app)
+    except Exception as e:
+        QMessageBox.critical(app, "Errore", f"Si è verificato un errore: {str(e)}")
+
+def update_person_availability(app, date, tipologia, fascia_oraria, dialog):
+    try:
+        person_name = app.person_list.currentItem().text()
+        if person_name not in app.person_schedule:
+            app.person_schedule[person_name] = {}
+
+        if date not in app.person_schedule[person_name]:
+            app.person_schedule[person_name][date] = []
+
+        app.person_schedule[person_name][date].append({
+            'tipologia': tipologia,
+            'fascia_oraria': fascia_oraria
+        })
+        
+        dialog.close()
+        display_person_details(app, app.person_list.currentItem())
         update_turni_table(app)
     except Exception as e:
         QMessageBox.critical(app, "Errore", f"Si è verificato un errore: {str(e)}")
@@ -39,6 +64,7 @@ def add_tipologia(app):
     new_tipologia, ok = QInputDialog.getText(app, 'Nuova Tipologia', 'Inserisci il nome della tipologia:')
     if ok and new_tipologia:
         app.tipologie_list.addItem(new_tipologia)
+        app.tipologia_schedule[new_tipologia] = {}
         update_week_display(app, new_tipologia)
 
 def remove_tipologia(app):
@@ -50,15 +76,16 @@ def remove_tipologia(app):
             if tipologia_name in app.tipologia_schedule:
                 del app.tipologia_schedule[tipologia_name]
             update_turni_table(app)
+            clear_week_display(app)
     except Exception as e:
         QMessageBox.critical(app, "Errore", f"Si è verificato un errore: {str(e)}")
 
 def update_turni_table(app):
     try:
-        # Clear existing layout
         if app.turni_table.layout():
-            for i in reversed(range(app.turni_table.layout().count())):
-                widget = app.turni_table.layout().itemAt(i).widget()
+            while app.turni_table.layout().count():
+                item = app.turni_table.layout().takeAt(0)
+                widget = item.widget()
                 if widget:
                     widget.deleteLater()
 
@@ -106,29 +133,35 @@ def save_fascia_oraria(app, tipologia, day, fascia_oraria):
     if day not in app.tipologia_schedule[tipologia]:
         app.tipologia_schedule[tipologia][day] = []
     app.tipologia_schedule[tipologia][day].append(fascia_oraria)
-    update_turni_table(app)
+    update_week_display(app, tipologia)
 
 def update_week_display(app, tipologia):
-    """Visualizza una settimana tipo (da lunedì a domenica) con possibilità di aggiungere fasce orarie."""
     clear_week_display(app)
     
     week_layout = QHBoxLayout()
-
-    # Giorni della settimana
     days_of_week = ["Lunedì", "Martedì", "Mercoledì", "Giovedì", "Venerdì", "Sabato", "Domenica"]
 
+    day_buttons = {}
     for day in days_of_week:
         day_button = QPushButton(day)
         day_button.setFixedSize(100, 100)
+        day_buttons[day] = day_button
         day_button.clicked.connect(lambda _, d=day: add_fascia_to_day(app, tipologia, d))
         week_layout.addWidget(day_button)
 
-    app.week_display.setLayout(week_layout)
-    app.week_display.update()
+    week_display_widget = QWidget()
+    week_display_widget.setLayout(week_layout)
+    app.week_display.setLayout(QVBoxLayout())
+    app.week_display.layout().addWidget(week_display_widget)
+
+    if tipologia in app.tipologia_schedule:
+        for day, fasce in app.tipologia_schedule[tipologia].items():
+            for button in day_buttons.values():
+                if button.text().startswith(day):
+                    button.setText(f"{day}\n" + "\n".join(fasce))
 
 def clear_week_display(app):
-    """Pulisce la visualizzazione della settimana."""
-    if app.week_display.layout():
+    if app.week_display.layout() is not None:
         while app.week_display.layout().count():
             item = app.week_display.layout().takeAt(0)
             widget = item.widget()
@@ -136,7 +169,7 @@ def clear_week_display(app):
                 widget.deleteLater()
 
 def add_fascia_to_day(app, tipologia, day):
-    """Apre un dialogo per aggiungere una fascia oraria al giorno selezionato."""
     fascia_oraria, ok = QInputDialog.getText(app, f'Aggiungi Fascia Oraria per {day}', 'Inserisci la fascia oraria:')
     if ok and fascia_oraria:
         save_fascia_oraria(app, tipologia, day, fascia_oraria)
+        update_week_display(app, tipologia)
